+++ packages/@guardian/source-foundations/src/typography/object-styles-to-string.ts
@@ -1,4 +1,4 @@
-import { TypographyStyles } from './types';
+import type { TypographyStyles } from './types';
 
 export const objectStylesToString = ({
 	fontFamily,
@@ -6,7 +6,7 @@ export const objectStylesToString = ({
 	lineHeight,
 	fontWeight,
 	fontStyle,
-}: TypographyStyles) => `
+}: TypographyStyles): string => `
 	font-family: ${fontFamily};
 	font-size: ${typeof fontSize === 'number' ? `${fontSize}px` : fontSize};
 	line-height: ${lineHeight};
